feat(cli): discourage using the argv override
