+++ detox/local-cli/test.js
@@ -11,6 +11,7 @@ const { composeDetoxConfig } = require('../src/configuration');
 const log = require('../src/utils/logger').child({ __filename });
 const { getPlatformSpecificString, printEnvironmentVariables } = require('./utils/misc');
 const { prependNodeModulesBinToPATH } = require('./utils/misc');
+const { DETOX_ARGV_OVERRIDE_NOTICE } = require('./utils/warnings');
 
 module.exports.command = 'test';
 module.exports.desc = 'Run your test suite with the test runner specified in package.json';
@@ -51,6 +52,8 @@ module.exports.handler = async function test(argv) {
 module.exports.middlewares = [
   function applyEnvironmentVariableAddendum(argv, yargs) {
     if (process.env.DETOX_ARGV_OVERRIDE) {
+      log.warn(DETOX_ARGV_OVERRIDE_NOTICE);
+
       return yargs.parse([
         ...process.argv.slice(2),
         ...parse(process.env.DETOX_ARGV_OVERRIDE),
